Updated for 1.0
Dormant population in old town buildings on populated maps is not counted for the editor requirement anymore.
Added a counter of people currently reading newspapers and people in the reach of the newspaper.
Fixed a bug where only people with higher education could read newspapers if the education difficulty was set to simple.
Fixed a bug where the maximum loyalty for newspapers was used also for schools instead of its own defined value.
The script should now properly take the game speed into account when calculating the increase in loyalty and culture.